requirements_dev.txt
Cython
numpy==1.23.5
scipy
librosa
pytorch-lightning==1.8.6
lightning-bolts==0.6.0.post1
#torch==1.13.1+cu117
pandas
matplotlib
mpi4py
scikit-learn==0.24.1
joblib
smplx
transforms3d
x-transformers
madmom
ftfy
axial_positional_embedding
einops
entmax
madgrad
loguru
mido
tianshou
timm
